chore: rename 'shopify' folder in data>graphql to 'storefront' / delete unused fragments.ts
