+++ app/routes/($locale).api.product.tsx
@@ -2,7 +2,7 @@ import {data as dataWithOptions} from '@shopify/remix-oxygen';
 import type {LoaderFunctionArgs} from '@shopify/remix-oxygen';
 
 import {getMetafields} from '~/lib/utils';
-import {PRODUCT_ITEM_QUERY} from '~/data/graphql/shopify/product';
+import {PRODUCT_ITEM_QUERY} from '~/data/graphql/storefront/product';
 
 export async function loader({request, context}: LoaderFunctionArgs) {
   const {storefront} = context;
